+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 1.5.0 (unreleased)
 ------------------
 
+- #112 Use default TZ when calculating birthdate if `on_date` param is not set
 - #111 Allow/Disallow the introduction of future dates of birth
 - #110 Compatibility with core#2584 (SampleType to DX)
 - #108 Fix Traceback when creating partitions when no patient assigned