+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 1.5.0 (unreleased)
 ------------------
 
+- #115 Fix non-latin names and MRNs comparison in a listing view
 - #113 Added estimated birthdate field to Patient content type
 - #112 Use default TZ when calculating birthdate if `on_date` param is not set
 - #111 Allow/Disallow the introduction of future dates of birth