+++ docs/changelog.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-1.5.0 (unreleased)
+1.5.0 (2025-04-04)
 ------------------
 
 - #123 Hyperlink patient on sample listing